docker-compose/services/backend/requirements.txt
fastapi==0.99.1
gunicorn
uvicorn
pyspark
pymongo
confluent-kafka
python-dotenv
jsons
requests
fastapi
pydantic
email-validator
pandas
load_dotenv
python-dotenv~=0.19.2
aiohttp
motor
pytz
redis